+++ code/libvorbis/lib/block.c
@@ -252,9 +252,9 @@ static int _vds_shared_init(vorbis_dsp_state *v,vorbis_info *vi,int encp){
   v->pcm=_ogg_malloc(vi->channels*sizeof(*v->pcm));
   v->pcmret=_ogg_malloc(vi->channels*sizeof(*v->pcmret));
   {
-    int i;
-    for(i=0;i<vi->channels;i++)
-      v->pcm[i]=_ogg_calloc(v->pcm_storage,sizeof(*v->pcm[i]));
+    int j;
+    for(j=0;j<vi->channels;j++)
+      v->pcm[j]=_ogg_calloc(v->pcm_storage,sizeof(*v->pcm[j]));
   }
 
   /* all 1 (large block) or 0 (small block) */
@@ -1025,9 +1025,9 @@ int vorbis_synthesis_lapout(vorbis_dsp_state *v,float ***pcm){
   }
 
   if(pcm){
-    int i;
-    for(i=0;i<vi->channels;i++)
-      v->pcmret[i]=v->pcm[i]+v->pcm_returned;
+    int k;
+    for(k=0;k<vi->channels;k++)
+      v->pcmret[k]=v->pcm[k]+v->pcm_returned;
     *pcm=v->pcmret;
   }