+++ code/libvorbis/lib/codebook.c
@@ -292,7 +292,7 @@ static ogg_uint32_t bitreverse(ogg_uint32_t x){
   return((x>> 1)&0x55555555) | ((x<< 1)&0xaaaaaaaa);
 }
 
-STIN long decode_packed_entry_number(codebook *book, oggpack_buffer *b){
+ long decode_packed_entry_number(codebook *book, oggpack_buffer *b){
   int  read=book->dec_maxlength;
   long lo,hi;
   long lok = oggpack_look(b,book->dec_firsttablen);