+++ code/renderervk/vk.c
@@ -274,9 +274,9 @@ static const qchar *vk_result_string( VkResult code ) {
 #undef CASE_STR
 
 #define VK_CHECK( function_call ) { \
-	VkResult res = function_call; \
-	if ( res < 0 ) { \
-		ri.Error( ERR_FATAL, "Vulkan: %s returned %s", #function_call, vk_result_string( res ) ); \
+	VkResult res1 = function_call; \
+	if ( res1 < 0 ) { \
+		ri.Error( ERR_FATAL, "Vulkan: %s returned %s", #function_call, vk_result_string( res1 ) ); \
 	} \
 }
 
@@ -3573,14 +3573,14 @@ static void vk_create_attachments( void )
 
 static void vk_create_framebuffers( void )
 {
-	VkImageView attachments[3];
+	VkImageView attachments1[3];
 	VkFramebufferCreateInfo desc;
 	uint32_t n;
 
 	desc.sType = VK_STRUCTURE_TYPE_FRAMEBUFFER_CREATE_INFO;
 	desc.pNext = NULL;
 	desc.flags = 0;
-	desc.pAttachments = attachments;
+	desc.pAttachments = attachments1;
 	desc.layers = 1;
 
 	for ( n = 0; n < vk.swapchain_image_count; n++ )
@@ -3591,8 +3591,8 @@ static void vk_create_framebuffers( void )
 		{
 			desc.width = gls.windowWidth;
 			desc.height = gls.windowHeight;
-			attachments[0] = vk.swapchain_image_views[n];
-			attachments[1] = vk.depth_image_view;
+			attachments1[0] = vk.swapchain_image_views[n];
+			attachments1[1] = vk.depth_image_view;
 			VK_CHECK( qvkCreateFramebuffer( vk.device, &desc, NULL, &vk.framebuffers.main[n] ) );
 
 			SET_OBJECT_NAME( vk.framebuffers.main[n], va( "framebuffer - main %i", n ), VK_DEBUG_REPORT_OBJECT_TYPE_FRAMEBUFFER_EXT );
@@ -3604,12 +3604,12 @@ static void vk_create_framebuffers( void )
 			{
 				desc.width = glConfig.vidWidth;
 				desc.height = glConfig.vidHeight;
-				attachments[0] = vk.color_image_view;
-				attachments[1] = vk.depth_image_view;
+				attachments1[0] = vk.color_image_view;
+				attachments1[1] = vk.depth_image_view;
 				if ( vk.msaaActive )
 				{
 					desc.attachmentCount = 3;
-					attachments[2] = vk.msaa_image_view;
+					attachments1[2] = vk.msaa_image_view;
 				}
 				VK_CHECK( qvkCreateFramebuffer( vk.device, &desc, NULL, &vk.framebuffers.main[n] ) );
 				SET_OBJECT_NAME( vk.framebuffers.main[n], "framebuffer - main", VK_DEBUG_REPORT_OBJECT_TYPE_FRAMEBUFFER_EXT );
@@ -3624,7 +3624,7 @@ static void vk_create_framebuffers( void )
 			desc.attachmentCount = 1;
 			desc.width = gls.windowWidth;
 			desc.height = gls.windowHeight;
-			attachments[0] = vk.swapchain_image_views[n];
+			attachments1[0] = vk.swapchain_image_views[n];
 			VK_CHECK( qvkCreateFramebuffer( vk.device, &desc, NULL, &vk.framebuffers.gamma[n] ) );
 
 			SET_OBJECT_NAME( vk.framebuffers.gamma[n], "framebuffer - gamma-correction", VK_DEBUG_REPORT_OBJECT_TYPE_FRAMEBUFFER_EXT );
@@ -3638,22 +3638,22 @@ static void vk_create_framebuffers( void )
 		desc.attachmentCount = 2;
 		desc.width = vk.screenMapWidth;
 		desc.height = vk.screenMapHeight;
-		attachments[0] = vk.screenMap.color_image_view;
-		attachments[1] = vk.screenMap.depth_image_view;
+		attachments1[0] = vk.screenMap.color_image_view;
+		attachments1[1] = vk.screenMap.depth_image_view;
 		if ( vk.screenMapSamples > VK_SAMPLE_COUNT_1_BIT )
 		{
 			desc.attachmentCount = 3;
-			attachments[2] = vk.screenMap.color_image_view_msaa;
+			attachments1[2] = vk.screenMap.color_image_view_msaa;
 		}
 		VK_CHECK( qvkCreateFramebuffer( vk.device, &desc, NULL, &vk.framebuffers.screenmap ) );
 		SET_OBJECT_NAME( vk.framebuffers.screenmap, "framebuffer - screenmap", VK_DEBUG_REPORT_OBJECT_TYPE_FRAMEBUFFER_EXT );
 
 		if ( vk.capture.image != VK_NULL_HANDLE )
 		{
-			attachments[0] = vk.capture.image_view;
+			attachments1[0] = vk.capture.image_view;
 
 			desc.renderPass = vk.render_pass.capture;
-			desc.pAttachments = attachments;
+			desc.pAttachments = attachments1;
 			desc.attachmentCount = 1;
 			desc.width = gls.captureWidth;
 			desc.height = gls.captureHeight;
@@ -3673,7 +3673,7 @@ static void vk_create_framebuffers( void )
 			desc.height = height;
 
 			desc.attachmentCount = 1;
-			attachments[0] = vk.bloom_image_view[0];
+			attachments1[0] = vk.bloom_image_view[0];
 
 			VK_CHECK( qvkCreateFramebuffer( vk.device, &desc, NULL, &vk.framebuffers.bloom_extract ) );
 
@@ -3690,10 +3690,10 @@ static void vk_create_framebuffers( void )
 
 				desc.attachmentCount = 1;
 
-				attachments[0] = vk.bloom_image_view[n+0+1];
+				attachments1[0] = vk.bloom_image_view[n+0+1];
 				VK_CHECK( qvkCreateFramebuffer( vk.device, &desc, NULL, &vk.framebuffers.blur[n+0] ) );
 
-				attachments[0] = vk.bloom_image_view[n+1+1];
+				attachments1[0] = vk.bloom_image_view[n+1+1];
 				VK_CHECK( qvkCreateFramebuffer( vk.device, &desc, NULL, &vk.framebuffers.blur[n+1] ) );
 
 				SET_OBJECT_NAME( vk.framebuffers.blur[n+0], va( "framebuffer - blur %i", n+0 ), VK_DEBUG_REPORT_OBJECT_TYPE_FRAMEBUFFER_EXT );
@@ -4002,25 +4002,25 @@ void vk_initialize( void )
 
 	// get memory size & defaults
 	{
-		VkPhysicalDeviceMemoryProperties props;
+		VkPhysicalDeviceMemoryProperties props1;
 		VkDeviceSize maxDedicatedSize = 0;
 		VkDeviceSize maxBARSize = 0;
-		qvkGetPhysicalDeviceMemoryProperties( vk.physical_device, &props );
-		for ( i = 0; i < props.memoryTypeCount; i++ ) {
-			if ( props.memoryTypes[i].propertyFlags == VK_MEMORY_HEAP_DEVICE_LOCAL_BIT ) {
-				maxDedicatedSize = props.memoryHeaps[props.memoryTypes[i].heapIndex].size;
+		qvkGetPhysicalDeviceMemoryProperties( vk.physical_device, &props1 );
+		for ( i = 0; i < props1.memoryTypeCount; i++ ) {
+			if ( props1.memoryTypes[i].propertyFlags == VK_MEMORY_HEAP_DEVICE_LOCAL_BIT ) {
+				maxDedicatedSize = props1.memoryHeaps[props1.memoryTypes[i].heapIndex].size;
 			}
-			else if ( props.memoryTypes[i].propertyFlags & VK_MEMORY_HEAP_DEVICE_LOCAL_BIT ) {
-				if ( maxDedicatedSize == 0 || props.memoryHeaps[props.memoryTypes[i].heapIndex].size > maxDedicatedSize ) {
-					maxDedicatedSize = props.memoryHeaps[props.memoryTypes[i].heapIndex].size;
+			else if ( props1.memoryTypes[i].propertyFlags & VK_MEMORY_HEAP_DEVICE_LOCAL_BIT ) {
+				if ( maxDedicatedSize == 0 || props1.memoryHeaps[props1.memoryTypes[i].heapIndex].size > maxDedicatedSize ) {
+					maxDedicatedSize = props1.memoryHeaps[props1.memoryTypes[i].heapIndex].size;
 				}
 			}
-			if ( props.memoryTypes[i].propertyFlags == (VK_MEMORY_PROPERTY_HOST_COHERENT_BIT | VK_MEMORY_PROPERTY_DEVICE_LOCAL_BIT) ) {
-				maxBARSize = props.memoryHeaps[props.memoryTypes[i].heapIndex].size;
+			if ( props1.memoryTypes[i].propertyFlags == (VK_MEMORY_PROPERTY_HOST_COHERENT_BIT | VK_MEMORY_PROPERTY_DEVICE_LOCAL_BIT) ) {
+				maxBARSize = props1.memoryHeaps[props1.memoryTypes[i].heapIndex].size;
 			}
-			else if ( (props.memoryTypes[i].propertyFlags & (VK_MEMORY_PROPERTY_HOST_COHERENT_BIT | VK_MEMORY_PROPERTY_DEVICE_LOCAL_BIT)) == (VK_MEMORY_PROPERTY_HOST_COHERENT_BIT | VK_MEMORY_PROPERTY_DEVICE_LOCAL_BIT) ) {
+			else if ( (props1.memoryTypes[i].propertyFlags & (VK_MEMORY_PROPERTY_HOST_COHERENT_BIT | VK_MEMORY_PROPERTY_DEVICE_LOCAL_BIT)) == (VK_MEMORY_PROPERTY_HOST_COHERENT_BIT | VK_MEMORY_PROPERTY_DEVICE_LOCAL_BIT) ) {
 				if ( maxBARSize == 0 ) {
-					maxBARSize = props.memoryHeaps[props.memoryTypes[i].heapIndex].size;
+					maxBARSize = props1.memoryHeaps[props1.memoryTypes[i].heapIndex].size;
 				}
 			}
 		}
@@ -4195,7 +4195,7 @@ void vk_initialize( void )
 	{
 		VkDescriptorPoolSize pool_size[3];
 		VkDescriptorPoolCreateInfo desc;
-		uint32_t i, maxSets;
+		uint32_t j, maxSets;
 
 		pool_size[0].type = VK_DESCRIPTOR_TYPE_COMBINED_IMAGE_SAMPLER;
 		pool_size[0].descriptorCount = MAX_DRAWIMAGES + 1 + 1 + 1 + VK_NUM_BLOOM_PASSES * 2; // color, screenmap, bloom descriptors
@@ -4209,8 +4209,8 @@ void vk_initialize( void )
 		pool_size[2].type = VK_DESCRIPTOR_TYPE_STORAGE_BUFFER_DYNAMIC;
 		pool_size[2].descriptorCount = 1;
 
-		for ( i = 0, maxSets = 0; i < ARRAY_LEN( pool_size ); i++ ) {
-			maxSets += pool_size[i].descriptorCount;
+		for ( j = 0, maxSets = 0; j < ARRAY_LEN( pool_size ); j++ ) {
+			maxSets += pool_size[j].descriptorCount;
 		}
 
 		desc.sType = VK_STRUCTURE_TYPE_DESCRIPTOR_POOL_CREATE_INFO;