+++ code/sdl/sdl_glimp.c
@@ -421,31 +421,31 @@ static qint GLW_SetMode( qint mode, const qchar *modeFS, qbool fullscreen, qbool
 
 		if ( fullscreen )
 		{
-			SDL_DisplayMode mode;
+			SDL_DisplayMode mode2;
 
 			switch ( testColorBits )
 			{
-				case 16: mode.format = SDL_PIXELFORMAT_RGB565; break;
-				case 24: mode.format = SDL_PIXELFORMAT_RGB24;  break;
+				case 16: mode2.format = SDL_PIXELFORMAT_RGB565; break;
+				case 24: mode2.format = SDL_PIXELFORMAT_RGB24;  break;
 				default: Com_DPrintf( "testColorBits is %d, can't fullscreen\n", testColorBits ); continue;
 			}
 
-			mode.w = config->vidWidth;
-			mode.h = config->vidHeight;
-			mode.refresh_rate = /* config->displayFrequency = */ Cvar_VariableIntegerValue( "r_displayRefresh" );
-			mode.driverdata = NULL;
+			mode2.w = config->vidWidth;
+			mode2.h = config->vidHeight;
+			mode2.refresh_rate = /* config->displayFrequency = */ Cvar_VariableIntegerValue( "r_displayRefresh" );
+			mode2.driverdata = NULL;
 
-			if ( SDL_SetWindowDisplayMode( SDL_window, &mode ) < 0 )
+			if ( SDL_SetWindowDisplayMode( SDL_window, &mode2 ) < 0 )
 			{
 				Com_DPrintf( "SDL_SetWindowDisplayMode failed: %s\n", SDL_GetError( ) );
 				continue;
 			}
 
-			if ( SDL_GetWindowDisplayMode( SDL_window, &mode ) >= 0 )
+			if ( SDL_GetWindowDisplayMode( SDL_window, &mode2 ) >= 0 )
 			{
-				config->displayFrequency = mode.refresh_rate;
-				config->vidWidth = mode.w;
-				config->vidHeight = mode.h;
+				config->displayFrequency = mode2.refresh_rate;
+				config->vidWidth = mode2.w;
+				config->vidHeight = mode2.h;
 			}
 		}